Move MetalExtension.
